preprocess_service/requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
requests
python-multipart
httpx
Pillow